requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
crispy-bootstrap5==0.7
Django==4.1.7
django-crispy-forms==2.0
djangorestframework==3.14.0
Pillow==9.4.0
psycopg2==2.9.5
pytz==2023.2
sqlparse==0.4.3
